requirements.txt
langchain
langchain-ollama
qdrant-client
beautifulsoup4
requests
streamlit
python-dotenv
numpy